requirements.txt
cx-Freeze==6.1
Pillow==7.0.0
